+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import Home from "./pages/home";
+import Home from "./pages";
 
 function App() {
   return (
