+++ src/pages/index.tsx
@@ -1,14 +1,14 @@
 import Header from "components/Header";
 import Footer from "components/Footer";
-import History from "pages/home/layouts/History";
-import HomePage from "pages/home/layouts/HomePage";
-import Service from "pages/service";
-import TimeLine from "pages/home/layouts/TimeLine";
-import Member from "pages/home/layouts/Member";
+import History from "pages/layouts/History";
+import HomePage from "pages/layouts/HomePage";
+import TimeLine from "pages/layouts/TimeLine";
+import Member from "pages/layouts/Member";
+import Service from "pages/layouts/Service";
 
 function Home() {
   return (
-    <div>
+    <>
       <Header />
       <HomePage />
       <History />
@@ -16,7 +16,7 @@ function Home() {
       <TimeLine />
       <Member />
       <Footer />
-    </div>
+    </>
   );
 }
